Move renderer.js into index.js
